Rearrange LocalAttestation code.
